move test to interpreter test
