+++ tests/dialects/test_memref.py
@@ -428,19 +428,3 @@ def test_memref_copy():
         match="Expected source and destination to have the same element type.",
     ):
         copy.verify()
-
-
-def test_memref_get_global():
-    memref_type = memref.MemRefType.from_element_type_and_shape(i64, (2, 2))
-    module = builtin.ModuleOp([])
-    with ImplicitBuilder(module.body):
-        define = memref.Global.get(
-            builtin.StringAttr("my_global"),
-            memref_type,
-            builtin.DenseArrayBase.from_list(i64, [1, 22, 3, 4]),
-            sym_visibility=builtin.StringAttr("public"),
-        )
-        with ImplicitBuilder(func.FuncOp("main", ((), ())).body):
-            fetch = memref.GetGlobal.get("my_global", memref_type)
-
-    assert SymbolTable.lookup_symbol(fetch, "my_global") is define
